+++ assert/user.go
@@ -70,9 +70,17 @@ func (a *userAssertion) End() {
 	a.assert(&a.s)
 }
 
+// TryOk returns true if the assertion passed.
+//
+// If asserted is false, the assertion was a no-op and ok is meaningless.
+func (a *userAssertion) TryOk() (ok bool, asserted bool) {
+	return a.s.skipped || !a.s.failed, true
+}
+
 // Ok returns true if the assertion passed.
 func (a *userAssertion) Ok() bool {
-	return a.s.skipped || !a.s.failed
+	ok, _ := a.TryOk()
+	return ok
 }
 
 // BuildReport generates a report about the assertion.