+++ test_test.go
@@ -110,7 +110,7 @@ var _ = Describe("type Test", func() {
 				It("can be called without making an assertion", func() {
 					test.AdvanceTime(
 						ByDuration(3*time.Second),
-						nil,
+						assert.Nothing,
 					)
 					Expect(t.Logs).To(ContainElement(
 						"--- ADVANCING TIME BY 3s ---",
@@ -138,7 +138,7 @@ var _ = Describe("type Test", func() {
 				It("can be called without making an assertion", func() {
 					test.AdvanceTime(
 						ToTime(time.Date(2100, 1, 2, 3, 4, 5, 6, time.UTC)),
-						nil,
+						assert.Nothing,
 					)
 					Expect(t.Logs).To(ContainElement(
 						"--- ADVANCING TIME TO 2100-01-02T03:04:05Z ---",
@@ -167,8 +167,10 @@ type noopAssertion struct{}
 
 func (noopAssertion) Begin(compare.Comparator) {}
 func (noopAssertion) End()                     {}
+func (noopAssertion) TryOk() (bool, bool)      { return true, true }
 func (noopAssertion) Ok() bool                 { return true }
 func (noopAssertion) Notify(fact.Fact)         {}
+
 func (noopAssertion) BuildReport(ok, verbose bool, r render.Renderer) *assert.Report {
 	return &assert.Report{
 		TreeOk:   ok,